+++ include/UserInterface.h
@@ -8,14 +8,14 @@
 class UserInterface
 {
     protected:
-        double wWidth;
-        double wHeight;
+        float wWidth;
+        float wHeight;
         int width;
         int height;
         int selectedEntityIndex;
 
     public:
-        UserInterface(double wWidth, double wHeight, int width = 0, int height = 0);
+        UserInterface(float wWidth, float wHeight, int width = 0, int height = 0);
 
         virtual void tick(Particle* entities, int nParticles, Particle* massiveParticles, int nPassiveParticles) = 0;
         virtual void drawParticle(Particle particle) = 0;
