+++ src/Vector.cpp
@@ -3,26 +3,26 @@
 #include <cstdlib>
 #include <iostream>
 
-//Vector::Vector(double magnitude, double direction) {
+//Vector::Vector(float magnitude, float direction) {
 //    this->magnitude = magnitude;
 //    this->direction = direction;
 //}
 
-Vector::Vector(double x, double y, double z) {
+Vector::Vector(float x, float y, float z) {
     this->x = x;
     this->y = y;
     this->z = z;
 }
 
-//void Vector::setMagnitude(double magnitude) {
+//void Vector::setMagnitude(float magnitude) {
 //    this->magnitude = magnitude;
 //}
 //
-//void Vector::setDirection(double direction) {
+//void Vector::setDirection(float direction) {
 //    this->direction = direction;
 //}
 
-void Vector::setComponents(double x, double y, double z) {
+void Vector::setComponents(float x, float y, float z) {
     //magnitude = sqrt(pow(x, 2.0) + pow(y, 2.0));
     //direction = atan2(y, x);
     this->x = x;
@@ -30,41 +30,41 @@ void Vector::setComponents(double x, double y, double z) {
     this->z = z;
 }
 
-double Vector::getMagnitude() {
+float Vector::getMagnitude() {
     //return magnitude;
     return sqrt(pow(x, 2.0) + pow(y, 2.0) + pow(z, 2.0));
 }
 //
-//double Vector::getDirection() {
+//float Vector::getDirection() {
 //    return direction;
 //}
 
-double Vector::getX() {
+float Vector::getX() {
     //return magnitude * cos(direction);
     return x;
 }
 
-double Vector::getY() {
+float Vector::getY() {
     //return magnitude * sin(direction);
     return y;
 }
 
-double Vector::getZ() {
+float Vector::getZ() {
     return z;
 }
 
 Vector Vector::normalize() {
-    double magnitude = getMagnitude();
+    float magnitude = getMagnitude();
 
     return Vector(x / magnitude, y / magnitude, z / magnitude);
 }
 
 Vector Vector::sum(Vector vector) {
-    //double x1 = getX();
-    //double y1 = getY();
-    double x2 = vector.getX();
-    double y2 = vector.getY();
-    double z2 = vector.getZ();
+    //float x1 = getX();
+    //float y1 = getY();
+    float x2 = vector.getX();
+    float y2 = vector.getY();
+    float z2 = vector.getZ();
 
     return Vector(x + x2, y + y2, z + z2);
 }
@@ -80,41 +80,41 @@ Vector Vector::operator+=(const Vector& vector) {
 }
 
 Vector Vector::difference(Vector vector) {
-    double x2 = vector.getX();
-    double y2 = vector.getY();
-    double z2 = vector.getZ();
+    float x2 = vector.getX();
+    float y2 = vector.getY();
+    float z2 = vector.getZ();
 
     return Vector(x - x2, y - y2, z - z2);
 }
 
-double Vector::dProduct(Vector vector) {
+float Vector::dProduct(Vector vector) {
 	return (x * vector.getX()) + (y * vector.getY()) + (z * vector.getZ());
 }
 
 Vector Vector::vProduct(Vector vector) {
-    double x2 = vector.getX();
-    double y2 = vector.getY();
-    double z2 = vector.getZ();
+    float x2 = vector.getX();
+    float y2 = vector.getY();
+    float z2 = vector.getZ();
 
     return Vector(y * z2 - y2 * z, z * x2 - z2 * x, x * y2 - x2 * y);
 }
 
 Vector Vector::orthogonal() {
-    double x2 = 0;
-    double y2 = 0;
-    double z2 = 0;
+    float x2 = 0;
+    float y2 = 0;
+    float z2 = 0;
 
     if (z != 0) {
-        x2 = (double) rand() / (double) RAND_MAX;
-        y2 = (double) rand() / (double) RAND_MAX;
+        x2 = (float) rand() / (float) RAND_MAX;
+        y2 = (float) rand() / (float) RAND_MAX;
         z2 = (-x * x2 - y * y2) / z;
     } else if (y != 0) {
-        x2 = (double) rand() / (double) RAND_MAX;
-        z2 = (double) rand() / (double) RAND_MAX;
+        x2 = (float) rand() / (float) RAND_MAX;
+        z2 = (float) rand() / (float) RAND_MAX;
         y2 = (-x * x2 - z * z2) / y;
     } else if (x != 0) {
-        z2 = (double) rand() / (double) RAND_MAX;
-        y2 = (double) rand() / (double) RAND_MAX;
+        z2 = (float) rand() / (float) RAND_MAX;
+        y2 = (float) rand() / (float) RAND_MAX;
         x2 = (-z * z2 - y * y2) / x;
     } 
 
@@ -123,6 +123,6 @@ Vector Vector::orthogonal() {
     return v.normalize();
 }
 
-Vector Vector::product(double scalar) {
+Vector Vector::product(float scalar) {
     return Vector(scalar * x, scalar * y, scalar * z);
 }
